requirements.txt
Django==1.4
django-bootstrap-toolkit==2.5.3
django-guardian==1.0.4
django-permissions==1.0.3
django-sitetree==0.7
django-smarter==0.4
django-taggit==0.9.3
easy-thumbnails==1.0-alpha-21